refactor: use cursor instead of offset to manage deep paging + better handle the log of the progress by encapsulating information into their own domain object

+++ src/main/java/uk/gov/nationalarchives/discovery/taxonomy/service/ProcessMessageService.java
@@ -1,8 +1,6 @@
 package uk.gov.nationalarchives.discovery.taxonomy.service;
 
 import org.apache.commons.lang3.StringUtils;
-import org.apache.solr.client.solrj.SolrClient;
-import org.apache.solr.client.solrj.SolrServerException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,11 +9,13 @@ import org.springframework.stereotype.Service;
 import uk.gov.nationalarchives.discovery.taxonomy.domain.exception.TaxonomyErrorType;
 import uk.gov.nationalarchives.discovery.taxonomy.domain.exception.TaxonomyException;
 import uk.gov.nationalarchives.discovery.taxonomy.domain.repository.Category;
+import uk.gov.nationalarchives.discovery.taxonomy.domain.repository.SearchQueryResultsWithCursor;
+import uk.gov.nationalarchives.discovery.taxonomy.domain.service.ProgressInformation;
 import uk.gov.nationalarchives.discovery.taxonomy.repository.CategoryRepository;
 import uk.gov.nationalarchives.discovery.taxonomy.repository.InformationAssetViewReadRepository;
+import uk.gov.nationalarchives.discovery.taxonomy.repository.InformationAssetViewWriteRepository;
 import uk.gov.nationalarchives.discovery.taxonomy.repository.UpdateRepository;
 
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -28,20 +28,18 @@ public class ProcessMessageService {
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
     private final CategoryRepository categoryRepository;
     private final InformationAssetViewReadRepository iaViewReadRepository;
+    private final InformationAssetViewWriteRepository informationAssetViewWriteRepository;
     private final UpdateRepository updateRepository;
-    private final SolrClient solrCloudReadServer;
-    private final SolrClient solrCloudWriteServer;
     @Value("${solr.cloud.read.query-page-size}")
     public Integer pageSize;
 
 
     @Autowired
-    public ProcessMessageService(CategoryRepository categoryRepository, InformationAssetViewReadRepository iaViewReadRepository, UpdateRepository updateRepository, SolrClient solrCloudReadServer, SolrClient solrCloudWriteServer) {
+    public ProcessMessageService(CategoryRepository categoryRepository, InformationAssetViewReadRepository iaViewReadRepository, InformationAssetViewWriteRepository informationAssetViewWriteRepository, UpdateRepository updateRepository) {
         this.categoryRepository = categoryRepository;
         this.iaViewReadRepository = iaViewReadRepository;
+        this.informationAssetViewWriteRepository = informationAssetViewWriteRepository;
         this.updateRepository = updateRepository;
-        this.solrCloudReadServer = solrCloudReadServer;
-        this.solrCloudWriteServer = solrCloudWriteServer;
     }
 
     public void publishCategory(String categoryId) {
@@ -54,23 +52,50 @@ public class ProcessMessageService {
         logger.info("Publishing category {}", category.getTtl());
         addCategoryToUncategorisedIAViews(category);
         removeCategoryFromCategorisedIAViews(category);
+
+        //FIXME since queries and updates are bounded, I could run updates as async tasks, rather than have a
+        // scheduled task, would make things look simpler
+        try {
+            while (updateRepository.hasPendingUpdates()) {
+                Thread.sleep(1000);
+            }
+            Thread.sleep(2000);
+            informationAssetViewWriteRepository.commit();
+        } catch (InterruptedException e) {
+            logger.error("process was interrupted", e);
+        }
     }
 
-    //TODO apply super dirty workaround on removal of categories too
     private void removeCategoryFromCategorisedIAViews(Category category) {
-        Integer offset = 0;
         Integer nbOfIaViews = iaViewReadRepository.countItemsNotMatchingQueryAndWithCategory(category.getQry(), category
                 .getCiaid(), category.getSc());
-        logger.info("removing category '{}' from {} IAVIews", category.getTtl(), nbOfIaViews);
-        int percentageOfProcessedItems=0;
-        while (offset < nbOfIaViews) {
-            List<String> iaids = iaViewReadRepository.searchItemsNotMatchingQueryAndWithCategory(category.getQry(),
-                    category.getCiaid(), hasThreshold(category.getSc()), offset, pageSize);
+        if (nbOfIaViews == 0) {
+            logger.info("category '{}' has no IAVIews to remove", category.getTtl());
+            return;
+        }
+        logger.info("removing category '{}' to {} IAVIews", category.getTtl(), nbOfIaViews);
+
+        ProgressInformation progressInformation = new ProgressInformation(nbOfIaViews);
+
+        String lastCursorMark = null;
+        while (true) {
+            logger.debug("processed {} IAVIews, cursor:{}", progressInformation.getNbOfProcessedItems(), lastCursorMark);
+            SearchQueryResultsWithCursor searchQueryResultsWithCursor = iaViewReadRepository.searchItemsNotMatchingQueryAndWithCategory(category.getQry(),
+                    category.getCiaid(), hasThreshold(category.getSc()), lastCursorMark, pageSize);
+
+            if (searchQueryResultsWithCursor.getCursor().equals(lastCursorMark)) {
+                logger.debug("completed updates, updated {}", progressInformation.getNbOfProcessedItems());
+                break;
+            }
+
+            List<String> iaids = searchQueryResultsWithCursor.getResults();
+            lastCursorMark = searchQueryResultsWithCursor.getCursor();
 
             updateRepository.removeCategoryFromIaids(category, iaids);
-            offset += pageSize;
-            percentageOfProcessedItems=logProgress(offset,nbOfIaViews, percentageOfProcessedItems);
+
+            updateAndLogProgress(progressInformation, iaids.size());
         }
+
         logger.info("finished removing category '{}'", category.getTtl());
     }
 
@@ -78,49 +103,45 @@ public class ProcessMessageService {
         return !(new Double(0).equals(score));
     }
 
-    private int logProgress(Integer offset, Integer totalNbOfItems, int lastPercentageOfProcessedItems) {
-        int percentageOfProcessedItems = offset*100/totalNbOfItems;
-        if (percentageOfProcessedItems>(lastPercentageOfProcessedItems+10) && percentageOfProcessedItems < 100){
-            logger.info("processed {} % so far",percentageOfProcessedItems);
-            return (percentageOfProcessedItems/10)*10;
+    private void updateAndLogProgress(ProgressInformation progressInformation, int newNbOfProcessedItems) {
+        progressInformation.addToNbOfProcessedItems(newNbOfProcessedItems);
+        int percentageOfProcessedItems = progressInformation.getNbOfProcessedItems() * 100 / progressInformation.getTotalNbOfIaViews();
+        if (percentageOfProcessedItems > (progressInformation.getPercentageOfProcessedItems() + 10) && percentageOfProcessedItems < 100) {
+            logger.info("processed {} % so far", percentageOfProcessedItems);
+            progressInformation.setPercentageOfProcessedItems((percentageOfProcessedItems / 10) * 10);
         }
-        return lastPercentageOfProcessedItems;
     }
 
-    //FIXME to clean and refactor
+    //FIXME duplicated code with removeCategoryFromCategorisedIAViews
     private void addCategoryToUncategorisedIAViews(Category category) {
         Integer nbOfIaViews = iaViewReadRepository.countItemsMatchingQueryAndWithoutCategory(category.getQry(), category
                 .getCiaid(), category.getSc());
-        if(nbOfIaViews==0){
+        if (nbOfIaViews == 0) {
             logger.info("category '{}' has no IAVIews to add", category.getTtl());
             return;
         }
         logger.info("adding category '{}' to {} IAVIews", category.getTtl(), nbOfIaViews);
-        int percentageOfProcessedItems=0;
-        do {
-            Integer offset = 0;
-            List<String> iaids = new ArrayList<>();
-            do {
-                logger.debug("processed {} IAVIews", offset);
-                iaids = iaViewReadRepository.searchItemsMatchingQueryAndWithoutCategory(category.getQry(),
-                        category.getCiaid(), hasThreshold(category.getSc()), offset, pageSize);
 
-                updateRepository.addCategoryToIaids(category, iaids);
-                offset += pageSize;
-                percentageOfProcessedItems = logProgress(offset, nbOfIaViews, percentageOfProcessedItems);
-            } while (iaids.size()>0);
-            try {
-                solrCloudWriteServer.commit();
-                Thread.sleep(2000);
-                solrCloudReadServer.commit();
-                Thread.sleep(2000);
-            } catch (SolrServerException | IOException e) {
-                e.printStackTrace();
-            } catch (InterruptedException e) {
-                e.printStackTrace();
+        ProgressInformation progressInformation = new ProgressInformation(nbOfIaViews);
+
+        String lastCursorMark = null;
+        while (true) {
+            logger.debug("processed {} IAVIews, cursor:{}", progressInformation.getNbOfProcessedItems(), lastCursorMark);
+            SearchQueryResultsWithCursor searchQueryResultsWithCursor = iaViewReadRepository.searchItemsMatchingQueryAndWithoutCategory(category.getQry(),
+                    category.getCiaid(), hasThreshold(category.getSc()), lastCursorMark, pageSize);
+
+            if (searchQueryResultsWithCursor.getCursor().equals(lastCursorMark)) {
+                logger.debug("completed updates, updated {}", progressInformation.getNbOfProcessedItems());
+                break;
             }
-        } while (iaViewReadRepository.countItemsMatchingQueryAndWithoutCategory(category.getQry(), category
-                .getCiaid(), category.getSc()) > 0);
+
+            List<String> iaids = searchQueryResultsWithCursor.getResults();
+            lastCursorMark = searchQueryResultsWithCursor.getCursor();
+
+            updateRepository.addCategoryToIaids(category, iaids);
+
+            updateAndLogProgress(progressInformation, iaids.size());
+        }
 
         logger.info("finished adding category '{}'", category.getTtl());
     }
